+++ sdk/react/src/Marker/Marker.type.js
@@ -1,0 +1,7 @@
+// @flow
+
+export type MarkerType = {
+  color?: string,
+  end?: number | 'end', // number of seconds or till the end
+  start: number | 'start', // number of seconds or from the start
+};
