+++ sdk/react/src/MarkersContainer/MarkersContainer.js
@@ -1,0 +1,35 @@
+// flow
+
+import * as React from 'react';
+import { StyleSheet, View } from 'react-native';
+
+import Marker, { type MarkerType } from '../Marker';
+import styles from './MarkersContainer.styles';
+
+type Props = {
+  duration: number,
+  markers: Array<MarkerType>,
+  style?: StyleSheet.Styles,
+};
+
+const MarkersContainer = ({ duration, markers, style }: Props) => (
+  markers.length > 0
+    ? (
+      <View style={[styles.root, style]}>
+        {markers.map((marker, index) => (
+          <Marker
+            duration={duration}
+            key={index} // eslint-disable-line react/no-array-index-key
+            marker={marker}
+          />
+        ))}
+      </View>
+    )
+    : null
+);
+
+MarkersContainer.defaultProps = {
+  style: undefined,
+};
+
+export default MarkersContainer;
