+++ sdk/react/src/MarkersContainer/index.js
@@ -1,0 +1,5 @@
+// @flow
+
+import MarkersContainer from './MarkersContainer';
+
+export default MarkersContainer;
