--- sdk/react/src/progressBar/Marker/Marker.type.js
@@ -1,7 +1,0 @@
-// @flow
-
-export type MarkerType = {
-  color?: string,
-  end?: number | 'end', // number of seconds or from the start
-  start: number | 'start', // number of seconds or till the end
-};
