--- sdk/react/src/progressBar/MarkersList/index.js
@@ -1,5 +1,0 @@
-// @flow
-
-import MarkersList from './MarkersList';
-
-export default MarkersList;
